blog post uploaded

+++ Week11/SpringBootProject1/TechTalentBlog/src/main/java/com/techtalentsouth/TechTalentBlog/BlogPost/BlogPostRepository.java
@@ -1,5 +1,7 @@
 package com.techtalentsouth.TechTalentBlog.BlogPost;
 
 import org.springframework.data.repository.CrudRepository;
+import java.math.BigInteger;
 
-public interface BlogPostRepository extends CrudRepository<BlogPost, Long>{}
+public interface BlogPostRepository extends CrudRepository<BlogPost, BigInteger>{
+}
